Make --force flag optional

+++ entrypoint.sh
@@ -68,8 +68,16 @@ git add .
 if git status | grep -q "Changes to be committed"
 then
   git commit --message "$INPUT_COMMIT_MESSAGE"
-  echo "Pushing git commit"
-  git push --force -u origin HEAD:"$OUTPUT_BRANCH"
+
+  if [ -z "$INPUT_USE_FORCE" ]
+  then
+    echo "Pushing git commit with --force"
+    git push --force -u origin HEAD:"$OUTPUT_BRANCH"
+  else
+    echo "Pushing git commit"
+    git push -u origin HEAD:"$OUTPUT_BRANCH"
+  fi
+
 else
   echo "No changes detected"
 fi
